+++ src/prototypes/time-control/time-control-03/_events/_hooks/index.tsx
@@ -1,12 +1,12 @@
-import { useResetAllListener } from './use-reset-all-listener'
+import { useResetAllEventListener } from '../_event-listener/use-reset-all-event-listener'
 
 /**
  * scope 全体のイベント購読をまとめて有効化する
  *
- * - 新しい購読 (`use-xxx-listener`) を追加する際は、ここに呼び出しを足すだけでよい
+ * - 新しい購読 (`use-xxx-event-listener`) を追加する際は、ここに呼び出しを足すだけでよい
  */
 export const ScopeEventListeners = () => {
-  useResetAllListener()
+  useResetAllEventListener()
 
   return null
 }
