+++ src/prototypes/time-control/time-control-03/_events/index.hooks.ts
@@ -1,2 +1,2 @@
-export * from './_hooks/_utils/use-time-control-03-event-dispatcher'
-export * from './_hooks/_utils/use-time-control-03-event-listener'
+export * from './_hooks/use-time-control-03-event-dispatcher'
+export * from './_hooks/use-time-control-03-event-listener'
